requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
gunicorn==23.0.0
whitenoise==6.8.2
psycopg2-binary==2.9.9
dj-database-url==2.2.0
requests==2.32.3
